clippy: fix partially 2

+++ src/mccs/src/daemon/engine.rs
@@ -95,7 +95,8 @@ impl DaemonEngine {
                     ProxyCompletion::InitCommunicator => {}
                     _ => panic!("unexpected result"),
                 };
-                let comm_handle = CommunicatorHandle((init.id as u64) << 32 + init.rank);
+                // TODO: whether is ((init.id as u64) << 32) + init.rank or (init.id as u64) << (32 + init.rank)
+                let comm_handle = CommunicatorHandle(((init.id as u64) << 32) + init.rank as u64);
                 let comm = CommunicatorDelegation {
                     comm_id: init.id,
                     cuda_device_idx: init.cuda_device_idx,
